requirements.txt
browser-use==0.5.7
fastapi==0.109.2
uvicorn==0.27.0.post1
python-multipart==0.0.9
MainContentExtractor==0.0.4
aiohttp==3.12.15
websockets==15.0.1
faiss-cpu==1.10.0
playwright==1.52.0
langchain>=0.3.0
langchain-openai>=0.3.0
langchain-anthropic>=0.3.0
langchain-core>=0.3.0
aiofiles>=24.1.0
anyio>=4.9.0
bubus>=1.4.5
google-api-core>=2.25.0
httpx>=0.28.1
markdownify==1.1.0
patchright>=1.52.5
portalocker>=2.7.0,<3.0.0
posthog>=3.7.0
psutil>=7.0.0
pydantic>=2.11.5
pyobjc>=11.0; platform_system == 'darwin'
pyperclip>=1.9.0
python-dotenv>=1.0.1
requests>=2.32.3
screeninfo>=0.8.1; platform_system != 'darwin'
typing-extensions>=4.12.2
uuid7>=0.1.0
authlib>=1.6.0
google-genai>=1.26.0
openai>=1.97.0
anthropic>=0.58.2
groq>=0.30.0
ollama>=0.5.1
google-api-python-client>=2.174.0
google-auth>=2.40.3
google-auth-oauthlib>=1.2.2
mcp>=1.10.1
pypdf>=5.7.0
markdown-pdf>=1.5